refactor: move prepared bundles out of sources
Move RecordBundle, SaveResult, and save_bundle into pipeline.prepare so sources owns traversal while prepare owns persistence-shaped write sets. This also removes the remaining bundle exports from the sources package and updates the focused persistence tests.\n\n🤖 Generated with [Claude Code](https://claude.com/claude-code)\n\nCo-Authored-By: Claude <noreply@anthropic.com>
