+++ tests/unit/pipeline/test_prepare.py
@@ -5,8 +5,8 @@ import json
 import pytest
 
 from polylogue.paths import is_within_root
+from polylogue.pipeline.prepare import RecordBundle, save_bundle
 from polylogue.rendering.renderers import HTMLRenderer
-from polylogue.sources import RecordBundle, save_bundle
 from polylogue.storage.backends.connection import open_connection
 from tests.infra.storage_records import make_attachment, make_conversation, make_message
